requirements.txt
Flask
gunicorn
google-analytics-data
pytrends
matplotlib